Update to CVES.cpp
Just a full slop update to the c++ version of CVES... python version so much better but really going to look to improve on this slop over time. Had to do quite a bit of manual debugging and it's still no-where near the vision I have for it... more work required. For now... just enjoy.